requirements.txt
pytest~=8.3.5
requests~=2.32.3
allure-pytest
allure-python-commons~=2.14.1